Added an error structure for the dictionary builder and  avoid panic in the function as much as possible.
