+++ lindera-ipadic-builder/Cargo.toml
@@ -18,11 +18,14 @@ categories = ["text-processing"]
 license = "MIT"
 
 [dependencies]
+anyhow = "1.0"
 bincode = "1.3"
 byteorder = "1.4"
 clap = "2.33"
 encoding = "0.2"
 glob = "0.3"
+serde = "1.0"
+thiserror = "1.0"
 yada = "0.4"
 
 lindera-core = { version = "0.7", path = "../lindera-core" }
